FileServer_Microservice/requirements.txt
boto3==1.28.73
ffmpeg_python==0.2.0
Flask==3.0.0
pydub==0.25.1
python-dotenv==1.0.0
